\name{CBpriceR}
\alias{CBpriceR}
\title{Calculating price of convertible bond}
\usage{
  CBpriceR(S_0, sigma, intRate0, creditSpread,
    rateCompounding, couponRate, couponFreq,
    dividendProtectionRate, RDividendSchedule,
    conversionRatio, noConversionPeriod,
    americanConversionType, maturity, faceValue,
    redemtionPremium, softCallStartTime, softCall,
    softCallStrike, putStartTime, putStrike, callStartTime,
    callStrike, nbStepsPerYear)
}
\arguments{
  \item{sigma}{volatility of the underlyng}
}
\value{
  price of the convertible
}
\description{
  This
}

