\name{convbond-package}
\alias{convbond-package}
\alias{convbond}
\docType{package}
\title{
Provide solution for pricing convertible bonds
}
\description{
Package provides framework for pricing convertible bonds based on Derman at al. binomial model. 
Also there are implemented some up to date features like dividend protection, embedded calls, puts etc.
}
\details{
\tabular{ll}{
Package: \tab convbond\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-07-30\cr
License: \tab >= GPU v2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Dmitriy Selivanov

Maintainer: Dmitriy Selivanov <dmitriy.selivanov@hmail.com>
}
\references{
http://www.ederman.com/new/docs/gs-valuing_convertibles.pdf
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
